+++ tests/database/test_db.py
@@ -1,20 +1,13 @@
-from pytest import MonkeyPatch
 from sqlalchemy.orm import Session
-from fastapi.testclient import TestClient
-
 import app.model as m
-import app.schema as s
-from app.controller import MailClient
 from tests.conftest import get_test_settings
 
 settings = get_test_settings()
 
 USER_NAME = "michael"
 USER_EMAIL = settings.TEST_TARGET_EMAIL
-USER_PASSWORD = "secret"
-USER_GOOGLE_ID = "123456789"
 USER_PICTURE_URL = "uploads/image.png"
-USER_ID = 12345
+USER_PASSWORD = "secret"
 
 
 def test_model_relations(db: Session, test_data: dict):
@@ -46,14 +39,8 @@ def test_model_relations(db: Session, test_data: dict):
     assert len(user_model.own_groups) == len(test_data["test_groups"])
 
 
-def test_forget_password(
-    client: TestClient,
-    db: Session,
-    mail_client: MailClient,
-    monkeypatch=MonkeyPatch,
-):
-
-    # creating user
+def test_create_user_subscription(db: Session):
+    # check if subscription belongs to users
     user = m.User(
         email=USER_EMAIL,
         username=USER_NAME,
@@ -62,41 +49,22 @@ def test_forget_password(
     )
     db.add(user)
     db.commit()
-    db.refresh(user)
 
-    # Testing if email has been sent
-    with mail_client.mail.record_messages() as outbox:
-        request = s.EmailSchema(email=USER_EMAIL)
-        response = client.post("/api/user/forgot_password", json=request.dict())
-        assert response.status_code == 200
-        assert len(outbox) == 1
-
-
-def test_create_user_subscription(db: Session):
-    # check if subscription belongs to users
-    user = m.User(
-        id=USER_ID,
-        email=USER_EMAIL,
-        username=USER_NAME,
-        password=USER_PASSWORD,
-        picture=USER_PICTURE_URL,
-    )
     subscription = m.Subscription(user_id=user.id)
-    db.add(user)
     db.add(subscription)
     db.commit()
     assert user.subscription
 
     # check if subscription DOES NOT belong to users
     user2 = m.User(
-        id=54321,
         email="non_existing_user@gmail.com",
         username="non_existing_username",
         password=USER_PASSWORD,
         picture=USER_PICTURE_URL,
     )
-    subscription2 = m.Subscription(user_id=999)
     db.add(user2)
+    db.commit()
+    subscription2 = m.Subscription(user_id=user.id)
     db.add(subscription2)
     db.commit()
     assert not user2.subscription
